+++ src/main/java/pooproject/udesc/poo/calculator/Validators.java
@@ -7,23 +7,23 @@ public class Validators {
     
     // Custom exceptions block
     private class InvalidDivider extends RuntimeException {
-        public InvalidDivider(Float number) {
+        public InvalidDivider(Integer number) {
             super(number.toString() + " is not a valid divider!");
         }
     }
     
     private class InvalidInputLimit extends RuntimeException {
-        public InvalidInputLimit(Float number) {
+        public InvalidInputLimit(Integer number) {
             super(number.toString() + " exceed the limit of 999!");
         }
     }
     
     // Validators
-    public void validateDivider(Float number) {
+    public void validateDivider(Integer number) {
         if (number == 0) throw new InvalidDivider(number);
     }
     
-    public void validateInputLimit(Float number) {
+    public void validateInputLimit(Integer number) {
         if (number >= 1000) throw new InvalidInputLimit(number);
     }
 }
